record every step
