+++ metadrive/base_class/base_object.py
@@ -244,9 +244,15 @@ class BaseObject(BaseRunnable):
     def set_static(self, flag):
         self.body.setStatic(flag)
 
+    def get_panda_pos(self):
+        return self.origin.getPos()
+
+    def set_panda_pos(self, pos):
+        self.origin.setPos(pos)
+
     def get_state(self) -> Dict:
         state = {
-            ObjectState.POSITION: self.position,
+            ObjectState.POSITION: self.get_panda_pos(),
             ObjectState.HEADING_THETA: self.heading_theta,
             ObjectState.ROLL: self.roll,
             ObjectState.PITCH: self.pitch,
@@ -255,7 +261,7 @@ class BaseObject(BaseRunnable):
         return state
 
     def set_state(self, state: Dict):
-        self.set_position(state[ObjectState.POSITION])
+        self.set_panda_pos(state[ObjectState.POSITION])
         self.set_heading_theta(state[ObjectState.HEADING_THETA])
         self.set_pitch(state[ObjectState.PITCH])
         self.set_roll(state[ObjectState.ROLL])
