+++ metadrive/manager/base_manager.py
@@ -26,7 +26,7 @@ class BaseManager(Randomizable):
         """
         pass
 
-    def after_step(self, *args, **kwargs) -> dict():
+    def after_step(self, *args, **kwargs) -> dict:
         """
         Update state for this manager after system advancing dt
         """
